th/requirements.txt
flask==2.2.5
flask-sqlalchemy
requests_toolbelt
kivy
kivymd
